finish publish

+++ publish.js
@@ -1,11 +1,17 @@
 import fs from "fs-extra";
 import { join } from "path";
+import esMain from "es-main";
 import { load } from "cheerio";
 import { execSync } from "child_process";
+import { LocalStorage } from "node-localstorage";
 
 const icons = fs.readJsonSync("./icons.json");
 const buildDir = join(process.cwd(), ".hfc", "build");
 const buildTemplateDir = join(process.cwd(), ".hfc", "build-tmpl");
+
+const localStorage = new LocalStorage("./tmp/localStorage");
+
+const publishFailed = [];
 
 let wfmHfcPath = "";
 async function publishIcon(icon) {
@@ -14,7 +20,15 @@ async function publishIcon(icon) {
     process.exit(-1);
   }
 
-  const { SVG_ATTRS, SVG_HTML } = await parseSvg(icon.path);
+  const publishKey = `${icon.name}-${icon.version}-done`;
+  const published = !!localStorage.getItem(publishKey);
+
+  if (published) {
+    console.log(icon.name, "already published");
+    return;
+  }
+
+  const { SVG_ATTRS, SVG_HTML } = parseSvg(icon.path);
 
   await fs.rm(buildDir, { recursive: true, force: true });
   await fs.copy(buildTemplateDir, buildDir, { recursive: true });
@@ -67,10 +81,20 @@ async function publishIcon(icon) {
   await fs.writeFile(wfmEntryPath, wfmEntry);
 
   await fs.copy(icon.path, join(process.cwd(), "banner.svg"));
+
+  const out = execSync("npm run publish-hfc -- --skip-build").toString();
+  if (out.includes("success")) {
+    console.log(icon.name, "published");
+    localStorage.setItem(publishKey, true);
+  } else {
+    console.log(out);
+    publishFailed.push(icon);
+    console.log(icon.name, "publish failed");
+  }
 }
 
-async function parseSvg(filePath) {
-  const svg = await fs.readFile(filePath, "utf8");
+export function parseSvg(filePath) {
+  const svg = fs.readFileSync(filePath, "utf8");
 
   const elem = load(svg)("svg");
   const attrs = elem.attr();
@@ -85,7 +109,7 @@ async function parseSvg(filePath) {
   return { SVG_ATTRS, SVG_HTML };
 }
 
-(async () => {
+async function run() {
   process.env.NODE_ENV = "production";
   await fs.rm(join(process.cwd(), ".hfc"), { recursive: true, force: true });
 
@@ -97,5 +121,14 @@ async function parseSvg(filePath) {
 
   for (const icon of icons) {
     await publishIcon(icon);
+    break;
   }
-})();
+
+  if (publishFailed.length) {
+    fs.writeFileSync("publish-failed.json", JSON.stringify(publishFailed));
+  }
+}
+
+if (esMain(import.meta)) {
+  run();
+}
